an_hook: switch to dynamic flag implementation
Summary:
Implement the new dynamic flag library, and replace an_hook.h with
a shim header that #define's the legacy an_hook interface in terms
of dynamic_flag.

Rather than control operators, like `AN_HOOK`, dynamic_flags exposes
macros that evaluate to the flag's value (0 or 1).

Dynamic flags macro are parameterised on two values:

1. The default value when the dynamic_flag library isn't initialised
   or otherwise fails to find the macro's hook point.
2. The initial value when the dynamic_flag library is initialised.

The `DF_FEATURE` macro evaluates to a feature flag that is false
by default, and false initially.  It can be enabled at runtime.

The `DF_DEFAULT` macro evaluates to a flag that is true by default
and true initially.  It can be disabled at runtime.

The `DF_OPT` macro evaluates to a flag that is true by default, and
false initially: this corresponds to code that is always safe to opt
into, but should ideally be enabled on demand.

The hook implementation is biased at compile-time: the path where we
fallthrough (hook disabled) is faster than the path where do jump to
a cold basic block (hook enabled).  That's exactly what we want for
`DF_FEATURE` and `DF_OPT`: we want the fast when the flag is disabled.

It's the opposite for `DF_DEFAULT`.  That's why we mark the hook point
as "flipped," so that the flag is enabled by making the hook point
fallthrough (unpatched) rather than jump (patched in).

Given the DF_ macros, which expand into expressions that evaluate to
0/1 integers, we can now implement all the AN_HOOK control operators.
This wasn't an option back when an_hook was initially written, because
the compilers at the time (gcc 4.8 and clang 4) couldn't be trusted to
propagate likely/unlikely through expressions like
`if (!__builtin_expect(...))`.  The only way to reliably get the
compilers to put the correct code in the slow path was to define
specialised macro for each case.  We don't have that problem anymore
(tested with GCC 7.3 and clang 9, the earliest clang with `asm goto`).

Depends on Dxxxx.

Minimally closes Txxxx.

Test Plan:
run and diff test/feature_flags (all 3 configs)
diff-objects test/feature_flags (all 3 configs)
godbolt with gcc 7 and clang 9

Reviewers: jshufro, ahicks, wandrews

Reviewed By: jshufro, wandrews

Maniphest Tasks: Txxxx

Differential Revision: https://foo.backtrace.io/Dxxxx