an_hook/dynamic_flag: add new DF_DEFAULT_SLOW dynamic flag kind
We sometimes want to enable code by default, but optimise the disabled
path.  The `DF_DEFAULT_SLOW` macro defines such a flag; it can be seen
as a version of `DF_DEFAULT` that optimises for the flag being false,
or a version of `DF_FEATURE` that is true by default and initially.

This makes sense for slower debug-only code that we might still want
to enable by default.

Part of Txxxx.

TESTED=ran feature_flags, diffed output.