dynamic_flags.c: correctly handle disjoint text segments
Before this commit, we'd end up always merging `mprotect` ranges,
which was not the intention (I believe the condition was reformulated
incorrectly a long time ago, in response to review feedback), but
happens to work for binaries that aren't linked with fancy flags.

We really want to keep merging `mprotect` ranges when the two ranges
are adjacent (or overlap), without any page between them: this way the
amortised `mprotect` call will only affect pages that would have been
`mprotect`ed aynyway.  With this commit, the amortisation condition
checks that the new `mprotect`ed range falls within +/- 1 page of the
old (accumulator) range.  Merging the new `mprotect` range with the
old range thus adds at most one page on each end of the old `mprotect`
range, so the extension can't be disjoint from the old range.

For example, if we know we already want to mprotect pages `[2, 4]`, we
can merge with a new record that would mprotect pages `[4, 5]`, or
`[1, 2]`: the resulting merged range (`[2, 5]` or `[1, 4]`) will only
include pages that would have been mprotected anyway.  Of course,
no-op merges are also safe.  That's how we get the new condition:
merge the new `mprotect` range in the old one if the former is within
one page (in either direction) of the latter.

In theory, this condition still misses some edge cases:

1. If the current range includes the zero page, `first_page - 1`
   will wrap around.  That's OK: unsigned overflow is
   defined, and it's not *incorrect* to amortise less
   than we could.

2. When the current patch instruction straddles two pages
   that are adjacent to the current range (e.g., current
   `mprotect` range is `[3, 10]`, and the current patch
   straddles pages `[11, 12]`), we could merge the range.
   That's a rare condition that may be hard to exercice,
   and, again, it's not *incorrect* to amortise less.
   Until there's a practical reason for merging such
   `mprotect` ranges, we'll live with the suboptimality.

We should now only err by missing amortisation opportunities. That's
unfortunate, but still correct.